fishnet_executor/requirements.txt
aleph-sdk-python==0.6.0
aars~=0.3.3
fishnet-cod~=0.1.0
semver